+++ .formatter.exs
@@ -1,5 +1,5 @@
 # Used by "mix format"
 [
   inputs: ["{mix,.formatter,.credo}.exs", "{config,lib,test}/**/*.{ex,exs}"],
-  import_deps: [:typedstruct]
+  import_deps: [:typedstruct, :peri]
 ]