+++ lib/atex/http/adapter.ex
@@ -2,9 +2,10 @@ defmodule Atex.HTTP.Adapter do
   @moduledoc """
   Behaviour for defining a HTTP client adapter to be used within atex.
   """
+  alias Atex.HTTP.Response
 
-  @type success() :: {:ok, map()}
-  @type error() :: {:error, integer(), map()} | {:error, term()}
+  @type success() :: {:ok, Response.t()}
+  @type error() :: {:error, Response.t() | term()}
   @type result() :: success() | error()
 
   @callback get(url :: String.t(), opts :: keyword()) :: result()